# Faster getTraits() methods