--- a/src/Reflection/Adapter/ReflectionClass.php
+++ b/src/Reflection/Adapter/ReflectionClass.php
@@ -384,14 +384,14 @@ final class ReflectionClass extends CoreReflectionClass
     {
         $traits = $this->betterReflectionClass->getTraits();
 
-        /** @var list<trait-string> $traitNames */
-        $traitNames = array_map(static fn (BetterReflectionClass $trait): string => $trait->getName(), $traits);
+        $traitsByName = [];
+        foreach ($traits as $trait) {
+            /** @var trait-string $traitName */
+            $traitName                = $trait->getName();
+            $traitsByName[$traitName] = new self($trait);
+        }
 
-        /** @psalm-suppress ImpureFunctionCall */
-        return array_combine(
-            $traitNames,
-            array_map(static fn (BetterReflectionClass $trait): self => new self($trait), $traits),
-        );
+        return $traitsByName;
     }
 
     /**