--- a/src/Reflection/Adapter/ReflectionEnum.php
+++ b/src/Reflection/Adapter/ReflectionEnum.php
@@ -312,14 +312,14 @@ final class ReflectionEnum extends CoreReflectionEnum
     {
         $traits = $this->betterReflectionEnum->getTraits();
 
-        /** @var list<trait-string> $traitNames */
-        $traitNames = array_map(static fn (BetterReflectionClass $trait): string => $trait->getName(), $traits);
+        $traitsByName = [];
+        foreach ($traits as $trait) {
+            /** @var trait-string $traitName */
+            $traitName                = $trait->getName();
+            $traitsByName[$traitName] = new ReflectionClass($trait);
+        }
 
-        /** @psalm-suppress ImpureFunctionCall */
-        return array_combine(
-            $traitNames,
-            array_map(static fn (BetterReflectionClass $trait): ReflectionClass => new ReflectionClass($trait), $traits),
-        );
+        return $traitsByName;
     }
 
     /** @return list<trait-string> */