5% faster in https://github.com/phpstan/phpstan/issues/13425