@ondrejmirtes 2025-09-04
Is it still 5% after https://github.com/phpstan/phpstan-src/pull/4268?