@staabm 2025-09-04
with this PR
```
➜  bug13425 git:(2.1.x) ✗ hyperfine '../../bin/phpstan analyze src/ plugins/' -i
Benchmark 1: ../../bin/phpstan analyze src/ plugins/
  Time (mean ± σ):     272.4 ms ±   1.7 ms    [User: 234.5 ms, System: 35.1 ms]
  Range (min … max):   269.6 ms … 275.0 ms    11 runs
 
  Warning: Ignoring non-zero exit code.
```

before this PR
```
➜  bug13425 git:(2.1.x) ✗ hyperfine '../../bin/phpstan analyze src/ plugins/' -i
Benchmark 1: ../../bin/phpstan analyze src/ plugins/
  Time (mean ± σ):     278.5 ms ±   3.4 ms    [User: 241.0 ms, System: 35.0 ms]
  Range (min … max):   274.5 ms … 286.7 ms    10 runs
 
  Warning: Ignoring non-zero exit code.
```

tested on latest macOS

```

 
➜  bug13425 git:(2.1.x) ✗ php -v
PHP 8.3.25 (cli) (built: Aug 26 2025 15:05:30) (NTS)
Copyright (c) The PHP Group
Zend Engine v4.3.25, Copyright (c) Zend Technologies
```

-> so its only 2% after https://github.com/phpstan/phpstan-src/pull/4268