@ondrejmirtes 2025-09-04
Yeah the absolutes of these percentages get much smaller once you solve the root cause and do much less in total 😊